Split project to parts
