+++ tests/models/guess_version_setuptools.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import sys
 from typing import NoReturn
 
-import pipdeptree
+from pipdeptree._models import guess_version
 
 if sys.version_info >= (3, 8):
     import importlib.metadata as importlib_metadata
@@ -16,4 +16,4 @@ def raise_import_error(name: str) -> NoReturn:
 
 
 importlib_metadata.version = raise_import_error  # type: ignore[assignment]
-print(pipdeptree.guess_version("setuptools"), end="")  # noqa: T201
+print(guess_version("setuptools"), end="")  # noqa: T201
