+++ tox.ini
@@ -27,7 +27,7 @@ set_env =
 commands =
     python -m pytest {tty:--color=yes} {posargs: \
       --junitxml {toxworkdir}{/}junit.{envname}.xml --cov {envsitepackagesdir}{/}pipdeptree \
-      --cov {toxinidir}{/}tests --cov-fail-under=75 \
+      --cov {toxinidir}{/}tests --cov-fail-under=85 \
       --cov-config=pyproject.toml --no-cov-on-fail --cov-report term-missing:skip-covered --cov-context=test \
       --cov-report html:{envtmpdir}{/}htmlcov --cov-report xml:{toxworkdir}{/}coverage.{envname}.xml \
       tests}
